feat: Use ES private properties instead of Symbols and defineProperty for privacy
BREAKING CHANGE: Due to dropping node 10.x support we use ES private properties instead of Symbols and `.defineProperty()`.

We've been using Symbols and `.defineProperty()` to define private properties in classes up until now. This has the downside that they are not true private and in later versions of node.js one would see these Symbols when inspecting an object. What we want is proper private properties.

This PR does also add a pretty printer which outputs an object literal or the object so when debugging one can see the getters and setters of the object.

Example: printing a object with `console.log()` would previously print the following:

```sh
PodiumClientResponse {
  [Symbol(podium:client:response:redirect)]: '',
  [Symbol(podium:client:response:content)]: '',
  [Symbol(podium:client:response:headers)]: {},
  [Symbol(podium:client:response:css)]: [],
  [Symbol(podium:client:response:js)]: []
}
```

Now the following will be printed:

```sh
{
  redirect: '',
  content: '',
  headers: {},
  css: [],
  js: []
}
```