fix(deps): remove clonedeep dependency
This replacement with a shallow clone might look odd, but only one
value at the root level is mutated. Also, because of a quirk with
how clonedeep works with the AssetCss and AssetJs classes in
podium utils, the entries in the css and js arrays weren't actually
cloned – instead the values were copied.

See https://github.com/podium-lib/utils/pull/289 for a more thorough
explanation with links to code.