feat: forward traceparent and tracestate headers (#474)
This makes it much easier to link OpenTelemetry traces between a
layout and its podlets. It's been possible before, but required
userland code to forward the headers in each fetch call. With
this, all users should need to get a distributed trace is to
insrument the layout and podlets with the OpenTelemetry SDK.

https://opentelemetry.io/docs/languages/js/getting-started/nodejs/